Initial commit: gRPC bindings & garbled circuits for Lighting trade PoC
Provide a garbled circuit implementation based on the paper:

  https://eprint.iacr.org/2014/756.pdf

This uses so-called half gates, to achieve the optimal number of two
ciphertexts per garbled bit, which in this implementation is 256 bits,
for 128-bit security. Half gates arise when either the circuit generator
or the evaluator knows one of the ungarbled input bits, and only require
one ciphertext per garbled bit. Each garbled circuit constructed (with
layout given by instances of the 'lightningtrade.cryptography.Program'
class) may contain a mixture of full or half gates.

The underlying cryptographic primitive used to build and evaluate each
garbled gate is a pseudorandom function (16-byte PRF, or hash), which is
based on AES-128 in this implementation. AES is chosen because it is
algebraically 'simple', which means that if a zero-knowledge proof (ZKP)
of circuit correctness (say based on zkSNARKs or zkSTARKs) is required,
it will hopefully be much smaller and cheaper to construct than if an
arbitrary fast PRF was used.

Also provide gRPC bindings to LND and some very preliminary code to make
Lightning transfers, as well as some very preliminary escrow protocol
logic using Java annotations.